refactor: 💡 rename trasnaction response model and decorator
also add IsURL validation to webhook and address other small PR comments
