+++ package.json
@@ -33,7 +33,7 @@
     "@nestjs/swagger": "^5.2.1",
     "@polymeshassociation/hashicorp-vault-signing-manager": "1.1.0",
     "@polymeshassociation/local-signing-manager": "1.2.0",
-    "@polymeshassociation/polymesh-sdk": "15.1.0-alpha.2",
+    "@polymeshassociation/polymesh-sdk": "^16.0.0",
     "@polymeshassociation/signing-manager-types": "1.0.0",
     "class-transformer": "0.5.1",
     "class-validator": "^0.13.2",
