+++ src/accounts/accounts.controller.ts
@@ -15,11 +15,11 @@ import { TransactionHistoryFiltersDto } from '~/accounts/dto/transaction-history
 import { TransferPolyxDto } from '~/accounts/dto/transfer-polyx.dto';
 import { PermissionsModel } from '~/accounts/models/permissions.model';
 import { BalanceModel } from '~/assets/models/balance.model';
-import { ApiArrayResponse, ApiCreatedOrSubscriptionResponse } from '~/common/decorators/swagger';
+import { ApiArrayResponse, ApiTransactionResponse } from '~/common/decorators/swagger';
 import { ExtrinsicModel } from '~/common/models/extrinsic.model';
 import { PaginatedResultsModel } from '~/common/models/paginated-results.model';
 import { TransactionQueueModel } from '~/common/models/transaction-queue.model';
-import { ApiTransactionResponse, handlePayload } from '~/common/utils';
+import { handlePayload, TransactionResponseModel } from '~/common/utils';
 
 @ApiTags('accounts')
 @Controller('accounts')
@@ -50,7 +50,7 @@ export class AccountsController {
     summary: 'Transfer an amount of POLYX to an account',
     description: 'This endpoint transfers an amount of POLYX to a specified Account',
   })
-  @ApiCreatedOrSubscriptionResponse({
+  @ApiTransactionResponse({
     description: 'Details about the transaction',
     type: TransactionQueueModel,
   })
@@ -63,7 +63,7 @@ export class AccountsController {
       '</ul>',
   })
   @Post('transfer')
-  async transferPolyx(@Body() params: TransferPolyxDto): Promise<ApiTransactionResponse> {
+  async transferPolyx(@Body() params: TransferPolyxDto): Promise<TransactionResponseModel> {
     const result = await this.accountsService.transferPolyx(params);
     return handlePayload(result);
   }
